Update README.md
This message showed up during the application process. As far as I understand that means that you cannot apply if you are not a citizen. Please correct me if I am wrong.

"All applicants must have or be eligible to obtain a security clearance. Several factors revolve around eligibility. As outlined in Executive Order 12968, Access to Classified Information, eligibility for access to classified information may only be granted to United States citizens. Are you a US Citizen?"